Update Register Screen (#15)
* Update Register Screen

Now, during registration, a user will need to input their username
(display name) and phone number as well. These data will be added into
the users collection in firebase.

Additionally, the FirebaseAuth object created upon creation of a new
account will have the displayName field updated with the input username.

* Remove the displayName from the home screen

There was issue with trying to retrieve the updated displayName after
updating the User object. An issue that we could possibly fix in the
future.